refactor paperscraper as keywords2papers so it can work well with papers2info, see the temp jup notebook, moved into auto-db-pipeline and __init__.py deleted for convenience
